requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-image-uploader-widget==1.0.0
django-jazzmin==3.0.0
django-phone-field==1.8.1
django-versatileimagefield==3.1
djangorestframework==3.15.2
drf-extensions==0.7.1
frozenlist==1.4.1
html5lib==1.1
idna==3.8
lxml==5.3.0
Markdown==3.7
multidict==6.0.5
numpy==2.1.0
oscrypto==1.3.0
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
twilio==9.2.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
yarl==1.9.7
psycopg2-binary==2.9.1
gunicorn==20.1.0
